refactor: extract carousel cards
