+++ shared/components/molecules/FeaturePages/CarouselCards/CarouselCards.types.ts
@@ -1,0 +1,7 @@
+export type CarouselCard = {
+    id: number
+    title: string
+    description: string,
+    image: string,
+    isActive: boolean
+}
