+++ shared/components/molecules/FeaturePages/CarouselCards/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './CarouselCards.view'
+export type { CarouselCard } from './CarouselCards.types'
